move common bits to a helper
